SmallString: Fix ambiguity of iterators
std::size_t is a undefined unsigned integer type and std::ptrdiff_t is a
undefined signed integer type. So sometimes the compiler doesn't know
which to choose because both could represent a integer.

+++ src/libs/utils/smallstring.h
@@ -267,12 +267,12 @@ public:
 
     reverse_iterator rbegin() noexcept
     {
-        return reverse_iterator(end() - 1l);
+        return reverse_iterator(end() - static_cast<std::size_t>(1));
     }
 
     reverse_iterator rend() noexcept
     {
-        return reverse_iterator(begin() - 1l);
+        return reverse_iterator(begin() - static_cast<std::size_t>(1));
     }
 
     const_iterator begin() const noexcept
