+++ src/libs/utils/smallstringview.h
@@ -106,12 +106,12 @@ public:
 
     const_reverse_iterator rbegin() const noexcept
     {
-        return const_reverse_iterator(end() - 1l);
+        return const_reverse_iterator(end() - static_cast<std::size_t>(1));
     }
 
     const_reverse_iterator rend() const noexcept
     {
-        return const_reverse_iterator(begin() - 1l);
+        return const_reverse_iterator(begin() - static_cast<std::size_t>(1));
     }
 
 private:
